Refactor common methods out of cmd and into core
