+++ lib/spack/spack/cmd/pkg.py
@@ -5,12 +5,9 @@
 
 from __future__ import print_function
 
-import os
-import re
 import sys
 
 import llnl.util.tty as tty
-from llnl.util.filesystem import working_dir
 from llnl.util.tty.colify import colify
 
 import spack.cmd
@@ -18,7 +15,6 @@ import spack.cmd.common.arguments as arguments
 import spack.paths
 import spack.repo
 import spack.util.package_hash as ph
-from spack.util.executable import which
 
 description = "query packages associated with particular git revisions"
 section = "developer"
@@ -82,78 +78,19 @@ def setup_parser(subparser):
     arguments.add_common_arguments(hash_parser, ['spec'])
 
 
-def packages_path():
-    """Get the test repo if it is active, otherwise the builtin repo."""
-    try:
-        return spack.repo.path.get_repo('builtin.mock').packages_path
-    except spack.repo.UnknownNamespaceError:
-        return spack.repo.path.get_repo('builtin').packages_path
-
-
-class GitExe:
-    # Wrapper around Executable for git to set working directory for all
-    # invocations.
-    #
-    # Not using -C as that is not supported for git < 1.8.5.
-    def __init__(self):
-        self._git_cmd = which('git', required=True)
-
-    def __call__(self, *args, **kwargs):
-        with working_dir(packages_path()):
-            return self._git_cmd(*args, **kwargs)
-
-
-_git = None
-
-
-def get_git():
-    """Get a git executable that runs *within* the packages path."""
-    global _git
-    if _git is None:
-        _git = GitExe()
-    return _git
-
-
-def list_packages(rev):
-    git = get_git()
-
-    # git ls-tree does not support ... merge-base syntax, so do it manually
-    if rev.endswith('...'):
-        ref = rev.replace('...', '')
-        rev = git('merge-base', ref, 'HEAD', output=str).strip()
-
-    output = git('ls-tree', '--name-only', rev, output=str)
-    return sorted(line for line in output.split('\n')
-                  if line and not line.startswith('.'))
-
-
 def pkg_add(args):
     """add a package to the git stage with `git add`"""
-    git = get_git()
-
-    for pkg_name in args.packages:
-        filename = spack.repo.path.filename_for_package_name(pkg_name)
-        if not os.path.isfile(filename):
-            tty.die("No such package: %s.  Path does not exist:" %
-                    pkg_name, filename)
-
-        git('add', filename)
+    spack.repo.add_package_to_git_stage(args.packages)
 
 
 def pkg_list(args):
     """list packages associated with a particular spack git revision"""
-    colify(list_packages(args.rev))
-
-
-def diff_packages(rev1, rev2):
-    p1 = set(list_packages(rev1))
-    p2 = set(list_packages(rev2))
-    return p1.difference(p2), p2.difference(p1)
+    colify(spack.repo.list_packages(args.rev))
 
 
 def pkg_diff(args):
     """compare packages available in two different git revisions"""
-    u1, u2 = diff_packages(args.rev1, args.rev2)
+    u1, u2 = spack.repo.diff_packages(args.rev1, args.rev2)
 
     if u1:
         print("%s:" % args.rev1)
@@ -168,64 +105,21 @@ def pkg_diff(args):
 
 def pkg_removed(args):
     """show packages removed since a commit"""
-    u1, u2 = diff_packages(args.rev1, args.rev2)
+    u1, u2 = spack.repo.diff_packages(args.rev1, args.rev2)
     if u1:
         colify(sorted(u1))
 
 
 def pkg_added(args):
     """show packages added since a commit"""
-    u1, u2 = diff_packages(args.rev1, args.rev2)
+    u1, u2 = spack.repo.diff_packages(args.rev1, args.rev2)
     if u2:
         colify(sorted(u2))
 
 
-def get_all_package_diffs(type, rev1='HEAD^1', rev2='HEAD'):
-    """Show packages changed, added, or removed (or any combination of those)
-       since a commit.
-
-    Arguments:
-
-        type (str): String containing one or more of 'A', 'B', 'C'
-        rev1 (str): Revision to compare against, default is 'HEAD^'
-        rev2 (str): Revision to compare to rev1, default is 'HEAD'
-
-    Returns:
-
-        A set contain names of affected packages.
-    """
-    lower_type = type.lower()
-    if not re.match('^[arc]*$', lower_type):
-        tty.die("Invald change type: '%s'." % type,
-                "Can contain only A (added), R (removed), or C (changed)")
-
-    removed, added = diff_packages(rev1, rev2)
-
-    git = get_git()
-    out = git('diff', '--relative', '--name-only', rev1, rev2,
-              output=str).strip()
-
-    lines = [] if not out else re.split(r'\s+', out)
-    changed = set()
-    for path in lines:
-        pkg_name, _, _ = path.partition(os.sep)
-        if pkg_name not in added and pkg_name not in removed:
-            changed.add(pkg_name)
-
-    packages = set()
-    if 'a' in lower_type:
-        packages |= added
-    if 'r' in lower_type:
-        packages |= removed
-    if 'c' in lower_type:
-        packages |= changed
-
-    return packages
-
-
 def pkg_changed(args):
     """show packages changed since a commit"""
-    packages = get_all_package_diffs(args.type, args.rev1, args.rev2)
+    packages = spack.repo.get_all_package_diffs(args.type, args.rev1, args.rev2)
 
     if packages:
         colify(sorted(packages))
