+++ var/spack/repos/builtin/packages/easi/package.py
@@ -8,7 +8,7 @@ import os
 import shutil
 
 from spack import *
-from spack.cmd.pkg import GitExe
+from spack.repo import GitExe
 
 
 class Easi(CMakePackage):
